feat: add `maxScore` prop & use it
