+++ src/song-card/song-card.css
@@ -92,6 +92,9 @@
 .dateAdded {
   font-size: 110%;
 }
+.maxScore {
+  font-size: 110%;
+}
 .folder {
   font-size: 75%;
 }
