+++ src/song-card/variants/ddr.css
@@ -8,7 +8,3 @@
   color: black;
   margin-top: -10px;
 }
-
-.exScore {
-  font-size: 110%;
-}
