+++ src/songs/ddr_world.json
@@ -79,6 +79,7 @@
       "difficult": "Difficult",
       "expert": "Expert",
       "challenge": "Challenge",
+      "maxScoreFormat": "EX: {maxScore}",
       "flareRank": "Flare skill rank unlocks",
       "unlock": "Unlockable songs",
       "goldExclusive": "20th Gold Cab Exclusive",
@@ -106,6 +107,7 @@
       "difficult": "Difficult",
       "expert": "Expert",
       "challenge": "Challenge",
+      "maxScoreFormat": "EX: {maxScore}",
       "flareRank": "フレアスキル解禁楽曲",
       "unlock": "解禁が必要な楽曲/譜面",
       "goldExclusive": "20th anniversary model(金筐体)限定",
